Update and rename create_entities.py to homeassistant_create_entities.py
- Added more detailed information on the script's functionality.
- Added the optional node_id, as discussed earlier with FransOV.
- Refined variable names for improved clarity.
- Renamed the file to clarify its purpose for Home Assistant.
